% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpPOP_nonlinear4.R
\name{SpPOP_nonlinear4}
\alias{SpPOP_nonlinear4}
\title{Generation of spatial population along with spatially varying model parameters generated based on non-linear function of latitudes and longitudes}
\usage{
SpPOP_nonlinear4(N, k, xlat, ylong)
}
\arguments{
\item{N}{integer; population size i.e. N= (k * k)}

\item{k}{integer}

\item{xlat}{numeric vector}

\item{ylong}{numeric vector}
}
\value{
The developed function returns a dataframe of spatially generated population consist of simulated response variable (i.e. Y) along with their spatial coordinates, spatially varying coefficients and one explanatory variable (i.e. X). The auxiliary variable
        has been generated following U(0,2) and the regression coefficients are generated as non-linear function of latitudes and longitudes.
}
\description{
Generation of spatial population along with spatially varying model parameters generated based on non-linear function of latitudes and longitudes
}
\examples{
sp_nonlinear4<-SpPOP_nonlinear4(100,10,c(1:10),c(1:10))
}
\references{
1. Wang, Ning., Mei, Chang-Lin. and Yan, Xiao-Dong. (2008). Local Linear Estimation of Spatially Varying Coefficient Models: An Improvement on the Geographically Weighted Regression Technique.
    Environment and Planning A: Economy and Space, 40(4), 986-1005.<DOI:10.1068/a3941>.


 2. Chao, Liu., Chuanhua, Wei. and Yunan, Su. (2018). Geographically weighted regression model-assisted estimation in survey sampling. Journal of Nonparametric Statistics. <DOI:10.1080/10485252.2018.1499907>.
}
